Missing Junit test and edited bug out were question could not be gained by its own id

+++ robo-app/src/main/java/org/softala/roboapp/repository/QuestionRepository.java
@@ -35,7 +35,7 @@ import org.springframework.stereotype.Repository;
  * 
  */
 @Repository
-public interface QuestionRepository extends CrudRepository<Question, Long> {
+public interface QuestionRepository extends CrudRepository<Question, Integer> {
 
 	@Query("select q from Question q where q.id = ?")
 	public Question findByQuestionId(String id);
